move EventType into a new module
